requirements.txt
Pillow~=8.3.2
numpy~=1.21.2
psd-tools~=1.9.18
opencv-python~=4.5.3.56
matplotlib~=3.4.3
